requirements-extractor-p1/requirements.txt
functions-framework==3.*
google-cloud-storage==3.3.0
google-cloud-firestore==2.21.0
google-cloud-pubsub==2.31.1
google-genai==1.32.0